+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.27.0
+
+### Features
+
+- Add nervos dao examples
+- Add smart contract examples
+
 ## 0.26.3
 
 ### Features